+++ nodejs/eks/cluster.ts
@@ -88,6 +88,20 @@ export interface CoreData {
     nodeSecurityGroupTags?: InputTags;
 }
 
+function createOrGetInstanceProfile(name: string, parent: pulumi.ComponentResource, instanceRoleName?: pulumi.Input<aws.iam.Role>, instanceProfileName?: pulumi.Input<string>): aws.iam.InstanceProfile {
+  let instanceProfile: aws.iam.InstanceProfile;
+  if (instanceProfileName) {
+    instanceProfile = aws.iam.InstanceProfile.get(`${name}-instanceProfile`,
+      instanceProfileName, undefined, {parent: parent});
+  } else {
+    instanceProfile = new aws.iam.InstanceProfile(`${name}-instanceProfile`, {
+        role: instanceRoleName,
+    }, { parent: parent });
+  }
+
+  return instanceProfile;
+}
+
 export function createCore(name: string, args: ClusterOptions, parent: pulumi.ComponentResource): CoreData {
     // Check to ensure that aws-iam-authenticator is installed, as we'll need it in order to deploy k8s resources
     // to the EKS cluster.
@@ -112,14 +126,19 @@ export function createCore(name: string, args: ClusterOptions, parent: pulumi.Co
     }
 
     // Create the EKS service role
-    const eksRole = new ServiceRole(`${name}-eksRole`, {
+    let eksRole: pulumi.Output<aws.iam.Role>;
+    if (args.serviceRole) {
+      eksRole = pulumi.output(args.serviceRole);
+    } else {
+      eksRole = (new ServiceRole(`${name}-eksRole`, {
         service: "eks.amazonaws.com",
         description: "Allows EKS to manage clusters on your behalf.",
         managedPolicyArns: [
             "arn:aws:iam::aws:policy/AmazonEKSClusterPolicy",
             "arn:aws:iam::aws:policy/AmazonEKSServicePolicy",
         ],
-    }, { parent: parent });
+      }, { parent: parent })).role;
+    }
 
     // Create the EKS cluster security group
     let eksClusterSecurityGroup: aws.ec2.SecurityGroup;
@@ -152,7 +171,7 @@ export function createCore(name: string, args: ClusterOptions, parent: pulumi.Co
 
     // Create the EKS cluster
     const eksCluster = new aws.eks.Cluster(`${name}-eksCluster`, {
-        roleArn: eksRole.role.apply(r => r.arn),
+        roleArn: eksRole.apply(r => r.arn),
         vpcConfig: {
             securityGroupIds: [eksClusterSecurityGroup.id],
             subnetIds: subnetIds,
@@ -255,9 +274,7 @@ export function createCore(name: string, args: ClusterOptions, parent: pulumi.Co
     } else if (args.instanceRole) {
         // Create an instance profile if using a default node group
         if (!args.skipDefaultNodeGroup) {
-            instanceProfile = new aws.iam.InstanceProfile(`${name}-instanceProfile`, {
-                role: args.instanceRole,
-            }, { parent: parent });
+            instanceProfile = createOrGetInstanceProfile(name, parent, args.instanceRole, args.instanceProfileName);
         }
 
         instanceRoleMappings = pulumi.output(args.instanceRole).apply(instanceRole =>
@@ -273,6 +290,7 @@ export function createCore(name: string, args: ClusterOptions, parent: pulumi.Co
                 "arn:aws:iam::aws:policy/AmazonEC2ContainerRegistryReadOnly",
             ],
         }, { parent: parent })).role;
+
         instanceRoles = pulumi.output([instanceRole]);
 
         // Create a new policy for the role, if specified.
@@ -285,9 +303,7 @@ export function createCore(name: string, args: ClusterOptions, parent: pulumi.Co
 
         // Create an instance profile if using a default node group
         if (!args.skipDefaultNodeGroup) {
-            instanceProfile = new aws.iam.InstanceProfile(`${name}-instanceProfile`, {
-                role: instanceRole,
-            }, { parent: parent });
+            instanceProfile = createOrGetInstanceProfile(name, parent, instanceRole, args.instanceProfileName);
         }
 
         instanceRoleMappings = pulumi.output(instanceRole).apply(role =>
@@ -408,6 +424,16 @@ export interface ClusterOptions {
      * Note: options `instanceRole` and `instanceRoles` are mutually exclusive.
      */
     instanceRole?: pulumi.Input<aws.iam.Role>;
+
+    /**
+     * The default IAM InstanceProfile to use on the Worker NodeGroups, if one is not already set in the NodeGroup.
+     */
+     instanceProfileName?: pulumi.Input<string>;
+
+    /**
+     * IAM Service Role for EKS to use to manage the cluster.
+     */
+    serviceRole?: pulumi.Input<aws.iam.Role>;
 
     /**
      * This enables the advanced case of registering *many* IAM instance roles